+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^6.26.1",
     "react-scripts": "5.0.1",
     "redux": "^5.0.1",
+    "sass": "^1.78.0",
     "typescript": "^5.0.3",
     "web-vitals": "^2.1.4"
   },
