+++ src/app/globalStyles/media.css
@@ -4,6 +4,8 @@
     --48-text: 36px;
     --72-text: 48px;
     --grid-5-2: 4;
+	  --grid-4-2: 3;
+
   }
   .footer__grid {
     grid-template-columns: 250px 1fr 250px;
@@ -101,6 +103,10 @@
   }
   .category-item__title {
     left: 20px;
+  }
+  .catalog {
+    grid-template-columns: 260px 1fr; 
+    gap: 40px;
   }
 }
 @media (max-width: 1199px) and (min-width: 0px) {
