+++ src/app/globalStyles/styles.css
@@ -648,18 +648,19 @@ gap: 20px;
 
 
 
-
-
-
-
-
-
 .popular-items__title {
 	margin-bottom: 24px;
 }
+
+
+.catalog {
+	display: grid;
+	grid-template-columns: 300px 1fr;
+	gap: 40px 80px;
+}
 .catalog__items {
 	display: grid;
-	grid-template-columns: repeat(var(--grid-5-2), 1fr);
+	grid-template-columns: repeat(var(--grid-4-2), 1fr);
 	gap: 25px;
 	transition: all 0.2s ease-out;
 }
@@ -668,6 +669,7 @@ gap: 20px;
 	flex-direction: column;
 	gap: 15px;
 }
+
 
 
 .catalog-item {
@@ -956,6 +958,38 @@ gap: 20px;
 	font-size: 12px;
 }
 
+
+
+
+.filters {
+    position: sticky;
+    top: 20px;
+    height: fit-content;
+}
+.filters__item {
+	display: flex;
+	flex-direction: column;
+	gap: 14px;
+	margin-bottom: 24px;
+
+}
+.filters__name {
+	font-size: var(--18-text);
+}
+.filters__range {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	gap: 10px;
+}
+.filters__checkboxs {
+	display: flex;
+	flex-direction: column;
+	gap: 6px;
+}
+
+
+
 .sort-price {
 	display: flex;
 	align-items: center;
@@ -964,6 +998,7 @@ gap: 20px;
 .sort-price__range {
 	display: flex;
 	flex-direction: column;
+	width: 100%;
 	background: var(--second-light-color);
 	border-radius: 16px;
 	padding: 8px 15px;
@@ -989,6 +1024,9 @@ gap: 20px;
 	background: #fff;
 	box-shadow: 0 0 0 0.2rem var(--accent-color-hover);
 }
+
+
+
