+++ src/components/CartItem/CartItem.module.scss
@@ -1,0 +1,34 @@
+.cart-item {
+
+	position: relative;
+	display: grid;
+	grid-template-columns: 60px 1fr 1fr;
+	grid-template-areas: "I T O";
+	align-items: center;
+	gap: 10px 20px;
+
+	padding: 10px 15px;
+	border: 4px solid var(--second-light-color);
+	border-radius: 24px;
+
+	&__title { font-size: var(--14-text);  grid-area: T;}
+
+	&__order {grid-area: O;
+		display: grid;
+		grid-template-columns: repeat(3, 120px) 40px;
+		align-items: center;
+		gap: 10px 20px;
+
+		.counter {width: 100px;}
+	}
+
+	img {grid-area: I;}
+
+	.del {
+		.icon { color: var(--second-color); }
+
+		&:hover {
+			.icon { color: var(--main-color-hover); }
+		}
+	}
+}
