+++ src/components/CatalogFilter/index.ts
@@ -1,0 +1,1 @@
+export { CatalogFilter } from "./CatalogFilter"
