+++ todo.md
@@ -3,17 +3,22 @@ TODO:
 [ ] : useSales сделать обычно функцией helpers
 [ ] : Автоматическая генерация роутинга
 [ ] : Query запросы для фильтрации
+[ ] : Сделать фильтры сбоку, сортировку сверху
+[ ] : Сделать появление фильтров frameMotion
+[ ] : Сделать смену блочного вида на строчный в каталоге
 [ ] : Регистрация/Авторизация
+[ ] : Сделать userSlice для хранения isAuth
+[ ] : Сделать хранение данных в localStorage
 [ ] : Слайдер для изображений в <Product />
 [x] : Доделать функционал добавления в корзину в <Product />
 [ ] : Сделать компонент для выпадающих элементов
 [x] : Реализовать Favorites
 [ ] : Реализовать Compare
 [ ] : Реализовать Breadcrumbs + page title
-[ ] : Сделать userSlice для хранения isAuth
 
 BAGS:
 [ ] : При переходе на страницу с query запросом, страница рендерится 5 раз
 [x] : Наложились стили .control
 [ ] : Иногда при переходах по категориям данные не успевают и вылетает ошибка
 [ ] : Инпут сортировки по цене можно вписать буквы
+[ ] : После выбора в CatalogFilters и обновлении страницы фильтры пропадают
